+++ genmonlib/mycommon.py
@@ -92,6 +92,22 @@ class MyCommon(object):
         z.update(y)
         return z
 
+    #---------------------MyCommon:urljoin--------------------------------------
+    def urljoin(self, *parts):
+        # first strip extra forward slashes (except http:// and the likes) and
+        # create list
+        part_list = []
+        for part in parts:
+            p = str(part)
+            if p.endswith('//'):
+                p = p[0:-1]
+            else:
+                p = p.strip('/')
+            part_list.append(p)
+        # join everything together
+        url = '/'.join(part_list)
+        return url
+
     #---------------------------------------------------------------------------
     def LogInfo(self, message, LogLine = False):